+++ src/lib/maths.ts
@@ -1,9 +1,8 @@
 export { default as Matrix } from "./maths/matrix.js";
 export { default as Matrix4 } from "./maths/matrix4.js";
-export * as M4H from "./maths/matrix4Helper.js";
+export * as Matrix4Func from "./maths/matrix4Func.js";
 export { default as Rectangle } from "./maths/rectangle.js";
-export * as Transforms from "./maths/transforms.js";
 export { default as Vector2 } from "./maths/vector2.js";
 export { default as Vector3 } from "./maths/vector3.js";
-export * as V3H from "./maths/vector3Helper.js";
+export * as Vector3Func from "./maths/vector3Func.js";
 export { default as Vector4 } from "./maths/vector4.js";
