+++ src/lib/maths/vector3Func.ts
